+++ projects/hslayers/src/components/data/catalogue/add-data-layer-descriptor.interface.ts
