+++ projects/hslayers/src/components/data/common/add-data-common.module.ts
@@ -4,9 +4,9 @@ import {FormsModule} from '@angular/forms';
 import {NgModule} from '@angular/core';
 import {TranslateModule} from '@ngx-translate/core';
 
-import {HsAddDataCommonVectorFileUploadDirective} from './addData-common-vector-file-upload.directive';
-import {HsAddDataTargetPositionComponent} from './addData-target-position.component';
-import {HsAddDataUrlComponent} from './addData-url.component';
+import {HsAddDataCommonVectorFileUploadDirective} from './add-data-common-vector-file-upload.directive';
+import {HsAddDataTargetPositionComponent} from './add-data-target-position.component';
+import {HsAddDataUrlComponent} from './add-data-url.component';
 import {HsGetCapabilitiesErrorComponent} from './capabilities-error-dialog.component';
 import {HsHistoryListModule} from '../../../common/history-list/history-list.module';
 import {HsNestedLayersTableComponent} from './nested-layers-table.component';
