+++ projects/hslayers/src/components/data/common/add-data-url.directive.html
