+++ projects/hslayers/src/components/data/file/add-data-file.module.ts
@@ -5,9 +5,9 @@ import {
 } from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
-import {HsAddDataFileComponent} from './addData-file.component';
-import {HsAddDataFileShpModule} from './shp/addData-file-shp.module';
-import {HsAddDataVectorModule} from '../vector/addData-vector.module';
+import {HsAddDataFileComponent} from './add-data-file.component';
+import {HsAddDataFileShpModule} from './shp/add-data-file-shp.module';
+import {HsAddDataVectorModule} from '../vector/add-data-vector.module';
 import {TranslateModule} from '@ngx-translate/core';
 
 @NgModule({
