+++ projects/hslayers/src/components/data/file/shp/add-data-file-layer.directive.html
