+++ projects/hslayers/src/components/data/file/shp/add-data-file-shp.component.ts
@@ -2,9 +2,9 @@ import BaseLayer from 'ol/layer/Base';
 import {Component, OnInit} from '@angular/core';
 
 import {FileDescriptor} from './file-descriptor.type';
-import {HsAddDataFileShpService} from './addData-file-shp.service';
+import {HsAddDataFileShpService} from './add-data-file-shp.service';
 
-import {HsAddDataUrlWmsService} from '../../url/wms/addData-url-wms.service';
+import {HsAddDataUrlWmsService} from '../../url/wms/add-data-url-wms.service';
 
 import {HsCommonEndpointsService} from '../../../../common/endpoints/endpoints.service';
 import {HsEndpoint} from '../../../../common/endpoints/endpoint.interface';
@@ -16,7 +16,7 @@ import {HsUtilsService} from '../../../utils/utils.service';
 
 @Component({
   selector: 'hs-add-data-file-shp',
-  templateUrl: './addData-file-layer.directive.html',
+  templateUrl: './add-data-file-layer.directive.html',
 })
 export class HsAddDataFileShpComponent implements OnInit {
   abstract: string;
