+++ projects/hslayers/src/components/data/file/shp/add-data-file-shp.module.ts
@@ -3,9 +3,9 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {TranslateModule} from '@ngx-translate/core';
 
-import {HsAddDataCommonModule} from '../../common/addData-common.module';
-import {HsAddDataFileShpComponent} from './addData-file-shp.component';
-import {HsAddDataFileShpService} from './addData-file-shp.service';
+import {HsAddDataCommonModule} from '../../common/add-data-common.module';
+import {HsAddDataFileShpComponent} from './add-data-file-shp.component';
+import {HsAddDataFileShpService} from './add-data-file-shp.service';
 import {HsUiExtensionsModule} from '../../../../common/widgets/ui-extensions.module';
 
 @NgModule({
