+++ projects/hslayers/src/components/data/url/add-data-url.html
