+++ projects/hslayers/src/components/data/url/arcgis/add-data-url-arcgis.directive.html
