+++ projects/hslayers/src/components/data/url/arcgis/add-data-url-arcgis.module.ts
@@ -3,9 +3,9 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {TranslateModule} from '@ngx-translate/core';
 
-import {HsAddDataArcGisComponent} from './addData-url-arcgis.component';
-import {HsAddDataArcGisService} from './addData-url-arcgis.service';
-import {HsAddDataCommonModule} from '../../common/addData-common.module';
+import {HsAddDataArcGisComponent} from './add-data-url-arcgis.component';
+import {HsAddDataArcGisService} from './add-data-url-arcgis.service';
+import {HsAddDataCommonModule} from '../../common/add-data-common.module';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
