+++ projects/hslayers/src/components/data/url/wms/add-data-url-wms.module.ts
@@ -3,9 +3,9 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {TranslateModule} from '@ngx-translate/core';
 
-import {HsAddDataCommonModule} from '../../common/addData-common.module';
-import {HsAddDataUrlWmsService} from './addData-url-wms.service';
-import {HsAddDataWmsComponent} from './addData-url-wms.component';
+import {HsAddDataCommonModule} from '../../common/add-data-common.module';
+import {HsAddDataUrlWmsService} from './add-data-url-wms.service';
+import {HsAddDataWmsComponent} from './add-data-url-wms.component';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
