+++ projects/hslayers/src/components/data/vector/add-data-vector.module.ts
@@ -3,10 +3,10 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {TranslateModule} from '@ngx-translate/core';
 
-import {HsAddDataCommonModule} from '../common/addData-common.module';
-import {HsAddDataVectorComponent} from './addData-vector.component';
-import {HsAddDataVectorService} from './addData-vector.service';
-import {HsVectorUrlParserService} from './addData-vector-url-parser.service';
+import {HsAddDataCommonModule} from '../common/add-data-common.module';
+import {HsAddDataVectorComponent} from './add-data-vector.component';
+import {HsAddDataVectorService} from './add-data-vector.service';
+import {HsVectorUrlParserService} from './add-data-vector-url-parser.service';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
